Move CachedCollection from lib to models
