+++ packages/rocketchat-models/client/index.js
@@ -2,10 +2,13 @@ import { Base } from './models/_Base';
 import Avatars from './models/Avatars';
 import Uploads from './models/Uploads';
 import UserDataFiles from './models/UserDataFiles';
+import { CachedCollection, CachedCollectionManager } from './models/CachedCollection';
 
 export {
 	Base,
 	Avatars,
 	Uploads,
 	UserDataFiles,
+	CachedCollection,
+	CachedCollectionManager,
 };
